# src/backend/src/GymTracker.Api/Contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Gym Tracker Admin Catalogs API
  version: 0.1.0
  description: Contrato para panel de administracion de musculos, mediciones y ejercicios.
servers:
  - url: http://localhost:5000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/admin/measurement-types/{id}:
    get:
      summary: Obtener medicion por id
      tags: [Admin Measurement Types]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Medicion encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementType'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar medicion
      tags: [Admin Measurement Types]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertMeasurementTypeRequest'
      responses:
        '200':
          description: Medicion actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementType'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      summary: Soft delete de medicion
      tags: [Admin Measurement Types]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Medicion borrada logicamente
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/measurement-types/{id}/reactivate:
    post:
      summary: Reactivar medicion
      tags: [Admin Measurement Types]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Medicion reactivada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeasurementType'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/measurement-types/import-csv:
    post:
      summary: Importar mediciones por CSV parseado
      tags: [Admin Measurement Types]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportMeasurementTypesRequest'
      responses:
        '200':
          description: Resultado de importacion parcial
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportMeasurementTypesResult'
...
  /api/admin/exercises/{id}:
    get:
      summary: Obtener ejercicio por id
      tags: [Admin Exercises]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Ejercicio encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Actualizar ejercicio
      tags: [Admin Exercises]
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertExerciseRequest'
      responses:
        '200':
          description: Ejercicio actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      summary: Soft delete de ejercicio
      tags: [Admin Exercises]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Ejercicio borrado logicamente
  /api/admin/exercises/{id}/reactivate:
    post:
      summary: Reactivar ejercicio
      tags: [Admin Exercises]
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Ejercicio reactivado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exercise'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/exercises/import-csv:
    post:
      summary: Importar ejercicios por CSV parseado
      tags: [Admin Exercises]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportExercisesRequest'
      responses:
        '200':
          description: Resultado de importacion parcial por fila
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportExercisesResult'
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Conflict:
      description: Conflicto de unicidad o regla de negocio
    NotFound:
      description: Recurso no encontrado
    ValidationError:
      description: Error de validacion de datos de entrada
  schemas:
...
    MeasurementType:
      type: object
      required: [id, key, name, unit, dataType, category, active, isDeleted, createdAt, updatedAt]
      properties:
        id: { type: string }
        key: { type: string }
        name: { type: string }
        unit: { type: string }
        dataType:
          type: string
          enum: [integer, decimal, time, boolean, text]
        category:
          type: string
          enum: [strength, cardio, mobility, general]
        description: { type: string, nullable: true }
        active: { type: boolean }
        isDeleted: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
...
    Exercise:
      type: object
      required: [id, code, name, difficulty, category, primaryMuscleIds, measurementTypeId]
      properties:
        id: { type: string }
        code: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        difficulty:
          type: string
          enum: [Beginner, Intermediate, Advanced]
        category:
          type: string
          enum: [Strength, Cardio, Mobility, Stretching]
        primaryMuscleIds:
          type: array
          items: { type: string }
        secondaryMuscleIds:
          type: array
          items: { type: string }
        measurementTypeId: { type: string }
        images:
          type: array
          items: { type: string }
        videos:
          type: array
          items: { type: string }
        deletedAt: { type: string, format: date-time, nullable: true }
...
    UpsertMeasurementTypeRequest:
      type: object
      required: [key, name, unit, dataType, category]
      properties:
        key: { type: string }
        name: { type: string }
        unit: { type: string }
        dataType:
          type: string
          enum: [integer, decimal, time, boolean, text]
        category:
          type: string
          enum: [strength, cardio, mobility, general]
        description: { type: string, nullable: true }
    UpsertExerciseRequest:
      type: object
      required: [code, name, primaryMuscleIds]
      description: |
        Canonical request uses category, difficulty and measurementTypeId.
        For backward compatibility, exerciseTypeId/exerciseTypeCode and formTypeId/formTypeCode are also accepted.
      anyOf:
        - required: [category, difficulty]
        - required: [exerciseTypeId, exerciseTypeCode]
      oneOf:
        - required: [measurementTypeId]
        - required: [formTypeId]
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        difficulty:
          type: string
          enum: [Beginner, Intermediate, Advanced]
        category:
          type: string
          enum: [Strength, Cardio, Mobility, Stretching]
        primaryMuscleIds:
          type: array
          minItems: 1
          items: { type: string }
        secondaryMuscleIds:
          type: array
          items: { type: string }
        measurementTypeId: { type: string }
        exerciseTypeId: { type: string }
        exerciseTypeCode: { type: string }
        formTypeId: { type: string }
        formTypeCode: { type: string }
        images:
          type: array
          items: { type: string }
        videos:
          type: array
          items: { type: string }
...
    ImportMeasurementTypesRequest:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ImportMeasurementTypeRow'
    ImportMeasurementTypeRow:
      type: object
      required: [key, name, unit, dataType, category]
      properties:
        key: { type: string }
        name: { type: string }
        unit: { type: string }
        dataType: { type: string }
        category: { type: string }
        description: { type: string, nullable: true }
    ImportMeasurementTypesResult:
      type: object
      required: [totalRows, createdRows, rejectedRows, rows]
      properties:
        totalRows: { type: integer, minimum: 0 }
        createdRows: { type: integer, minimum: 0 }
        rejectedRows: { type: integer, minimum: 0 }
        rows:
          type: array
          items:
            $ref: '#/components/schemas/ImportMeasurementTypeRowResult'
    ImportExercisesRequest:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ImportExerciseRow'
    ImportExerciseRow:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        category: { type: string }
        difficulty: { type: string }
        measurementTypeId: { type: string }
        formTypeId: { type: string }
        formTypeCode: { type: string }
        primaryMuscleIds:
          type: array
          items: { type: string }
        secondaryMuscleIds:
          type: array
          items: { type: string }
    ImportExercisesResult:
      type: object
      required: [totalRows, createdRows, rejectedRows, rows]
      properties:
        totalRows: { type: integer }
        createdRows: { type: integer }
        rejectedRows: { type: integer }
        rows:
          type: array
          items:
            $ref: '#/components/schemas/ImportExerciseRowResult'
    ImportExerciseRowResult:
      type: object
      required: [rowNumber, created]
      properties:
        rowNumber: { type: integer }
        code: { type: string }
        created: { type: boolean }
        reason: { type: string, nullable: true }
        exercise:
          allOf:
            - $ref: '#/components/schemas/Exercise'
          nullable: true
    ImportMeasurementTypeRowResult:
      type: object
      required: [rowNumber, created]
      properties:
        rowNumber: { type: integer, minimum: 1 }
        key: { type: string, nullable: true }
        created: { type: boolean }
        reason: { type: string, nullable: true }
        measurementType:
          oneOf:
            - $ref: '#/components/schemas/MeasurementType'
            - type: 'null'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT